Add decoration classes
